+++ tests_for_fem/elasticity2d_code.py
@@ -1,0 +1,44 @@
+"For testing purposes only"
+
+from jax_am.fem.core import FEM
+import jax.numpy as np
+import jax
+
+
+class Elasticity(FEM):
+    def custom_init(self):
+        """Override base class method.
+        Modify self.flex_inds so that location-specific TO can be realized.
+        """
+        self.flex_inds = np.arange(len(self.cells))
+
+    def get_tensor_map(self):
+        """Override base class method.
+        """
+        def stress(u_grad, theta):
+            # Plane stress assumption
+            # Reference: https://en.wikipedia.org/wiki/Hooke%27s_law
+            Emax = 70.e3
+            Emin = 1e-3*Emax
+            nu = 0.3
+            penal = 3.
+            E = Emin + (Emax - Emin)*theta[0]**penal
+            epsilon = 0.5*(u_grad + u_grad.T)
+            eps11 = epsilon[0, 0]
+            eps22 = epsilon[1, 1]
+            eps12 = epsilon[0, 1]
+            sig11 = E/(1 + nu)/(1 - nu)*(eps11 + nu*eps22)
+            sig22 = E/(1 + nu)/(1 - nu)*(nu*eps11 + eps22)
+            sig12 = E/(1 + nu)*eps12
+            sigma = np.array([[sig11, sig12], [sig12, sig22]])
+            return sigma
+        return stress
+
+    def set_params(self, params):
+        """Override base class method.
+        """
+        full_params = np.ones((self.num_cells, params.shape[1]))
+        full_params = full_params.at[self.flex_inds].set(params)
+        thetas = np.repeat(full_params[:, None, :], self.num_quads, axis=1)
+        self.full_params = full_params
+        self.internal_vars['laplace'] = [thetas]
